requirements.txt
aiogram==3.7.0
python-dotenv==1.0.1
aiosqlite==0.20.0
APScheduler==3.10.4
astral==3.2
Pillow==10.4.0
geopy==2.4.1
pytz==2024.1
